hexahedron: (i386/x86_64:interrupt) Take uintptr_t instead of uint32_t

+++ hexahedron/include/kernel/arch/i386/interrupt.h
@@ -53,8 +53,8 @@ typedef struct _i386_gdtr {
 } i386_gdtr_t;
 
 // Interrupt/exception handlers
-typedef int (*interrupt_handler_t)(uint32_t, uint32_t, registers_t*, extended_registers_t*);
-typedef int (*exception_handler_t)(uint32_t, registers_t*, extended_registers_t*);
+typedef int (*interrupt_handler_t)(uintptr_t, uintptr_t, registers_t*, extended_registers_t*);
+typedef int (*exception_handler_t)(uintptr_t, registers_t*, extended_registers_t*);
 
 /**** DEFINITIONS ****/
